requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
gate-api==4.60.2